+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file. This projec
 
 ## 4.0.4-dev
 
+### Fixed
+
+- Fixed return type for `local:commitAndPush` causing exception (#2723)
 ## 4.0.3 - 2025-09-11
 
 ### Added